+++ programs/gmsol-store/src/instructions/oracle/custom.rs
@@ -3,7 +3,7 @@ use chainlink_datastreams::interface::ChainlinkDataStreamsInterface;
 use gmsol_utils::InitSpace;
 
 use crate::{
-    states::{PriceFeed, PriceFeedPrice, PriceProviderKind, Seed, Store},
+    states::{AmountKey, PriceFeed, PriceFeedPrice, PriceProviderKind, Seed, Store},
     utils::internal,
     CoreError,
 };
@@ -111,7 +111,13 @@ pub(crate) fn unchecked_update_price_feed_with_chainlink(
     let price = accounts.decode_and_validate_report(&compressed_report)?;
     accounts.verify_report(compressed_report)?;
 
-    accounts.price_feed.load_mut()?.update(&price)?;
+    accounts.price_feed.load_mut()?.update(
+        &price,
+        *accounts
+            .store
+            .load()?
+            .get_amount_by_key(AmountKey::OracleMaxFutureTimestampExcess),
+    )?;
 
     Ok(())
 }